라인 저장 & 로드 기능 추가
- 라인 저장 및 로드 기능 추가
- 블록 ID 설정하는 부분에 문제 있음 수정해야 함

+++ Assets/Scripts/REEL.EAIEditor/Manager/KeyInputManager.cs
@@ -7,36 +7,29 @@ namespace REEL.EAIEditor
     public class KeyInputManager : Singleton<KeyInputManager>
     {
         public bool shouldMultiSelect = false;
+        public bool isShiftPressed = false;
+        public bool isAltPressed = false;
 
         void Update()
         {
 #if UNITY_EDITOR
-            if (Input.GetKey(KeyCode.LeftShift) && Input.GetKeyDown(KeyCode.A))
-            {
-                BlockDiagramManager.Instance.SetAllSelected();
-            }
+            if (Input.GetKey(KeyCode.LeftShift) && Input.GetKeyDown(KeyCode.A)) BlockDiagramManager.Instance.SetAllSelected();
 
 #else
-            if (shouldMultiSelect && Input.GetKeyDown(KeyCode.A))
-            {
-                BlockDiagramManager.Instance.SetAllSelected();
-            }
+            if (shouldMultiSelect && Input.GetKeyDown(KeyCode.A)) BlockDiagramManager.Instance.SetAllSelected();
 #endif
 
-            if (Input.GetKeyDown(KeyCode.Delete))
-            {
-                BlockDiagramManager.Instance.DeleteSelected();
-            }
+            if (Input.GetKeyDown(KeyCode.Delete)) BlockDiagramManager.Instance.DeleteSelected();
+            if (Input.GetKeyDown(KeyCode.Escape)) BlockDiagramManager.Instance.SetAllUnselected();
 
-            if (Input.GetKeyDown(KeyCode.LeftControl))
-            {
-                shouldMultiSelect = true;
-            }
+            if (Input.GetKeyDown(KeyCode.LeftControl)) shouldMultiSelect = true;
+            if (Input.GetKeyUp(KeyCode.LeftControl)) shouldMultiSelect = false;
 
-            if (Input.GetKeyUp(KeyCode.LeftControl))
-            {
-                shouldMultiSelect = false;
-            }
+            if (Input.GetKeyDown(KeyCode.LeftShift) || Input.GetKeyDown(KeyCode.RightShift)) isShiftPressed = true;
+            if (Input.GetKeyUp(KeyCode.LeftShift) || Input.GetKeyDown(KeyCode.RightShift)) isShiftPressed = false;
+
+            if (Input.GetKeyDown(KeyCode.LeftAlt)) isAltPressed = true;
+            if (Input.GetKeyUp(KeyCode.LeftAlt)) isAltPressed = false;
         }
     }
 }
